+++ test/notification_test.exs
@@ -1,7 +1,7 @@
 defmodule NotificationTest do
   use ExUnit.Case, async: true
   import Postgrex.TestHelper
-  alias Postgrex.Connection, as: P
+  alias Postgrex, as: P
   alias Postgrex.Notifications, as: PN
 
   setup do
